+++ CHANGELOG.md
@@ -27,8 +27,10 @@ v0.19
 * Adds ``Collection.add_batches`` and ``Collection.update_batches`` which are variants on
   ``Collection.add`` and ``Collection.update`` that will return generators
   instead of adding/updating everything in-memory.
-* The ``Collection.update`` and ``Collection.update_batches`` will check for equality between ``Documents`` before updating.
-This makes it possible to skip updates in particular cases by overriding ``Document.__eq__``.
+* The ``Collection.update``, ``Collection.add``, ``Collection.update_batches`` and ``Collection.add_batches`` will
+  check for equality between ``Documents`` before adding or updating. This makes it possible to skip insert/updates in
+  particular cases by overriding ``Document.__eq__``. ``Collection.add`` and ``Collection.add_batches`` require
+  input as a list for this to work to prevent unexpected excessive memory usage.
 * Uses ``Collection.document_update_fields`` to determine which fields to update in ``bulk_update`` calls by Collection.
 * Adds ``Document.build`` to support creating a ``Document`` from raw data.
 * ``Document.update`` will now use properties as update data instead of content